update Puppet file
this commit contains a small refactor of the Puppetfile.

Adds an environment variable git_protocol, that can be used
to update the run from git to https for proxzy related issues.

Starts to add a skeleton of the code that will be used to specify
that installs can either happen based on upstream or downstream.